requirements.txt
Flask==2.3.4
Flask-SQLAlchemy==3.0.2
Flask-Login==0.6.2
Flask-WTF==1.0.1
Flask-Migrate==4.0.4
Flask-Cors==3.2.1
geopy==2.3.0
matplotlib==3.8.0
Pillow==9.5.0
opencv-python==4.8.0.74
Werkzeug==2.3.6
SQLAlchemy==2.1.0
tensorflow==2.14.0
authlib==1.0.0
scikit-learn==1.2.2
numpy==1.24.3
wtforms==3.0.1
face-recognition==1.3.0
Flask-Dance==7.1.0
oauthlib==3.2.2
requests-oauthlib==2.0.0
email-validator==2.2.0
pillow-heif==0.18.0
